Update patch set 4
Patch Set 4:

> Michael, this looks pretty cool. Much less error-prone.

Thanks! Stephen and I are both excited about taking things into this direction.

 > Are there annotations for finding different versions of services?
 > For example, the DataBroker has a few different impls, of which we
 > use the odl-default, but there are other's like ping-pong.

That's a good point.  My understanding is that we could either continue to use BP XML (mixed, for now) or use the @OsgiService annotation which "creates a reference to an OSGi service" - given a name (String).

I'm not familiar with alternative DataBroker implementation.  Can you point me to a project which uses ping-ping instead of the default, so that we can look at that wiring XML, and see how we would do it here?

 > There are also notification providers for things like the
 > packetHandlers where the service implements an interface.

Can you clarify? Best again with a link to a project using this.  If you mean the custom ODL BP XML elements, then our idea for the short term is that this would stay in XML (it's possible to mix).  Getting already the "trivial" bean wiring out XML would be a valuable step worth doing.

Patch-set: 4
Label: Verified=0