Update patch set 4
Patch Set 4:

> <reference id="dataBroker"
 > interface="org.opendaylight.controller.md.sal.dom.api.DOMDataBroker"
 > odl:type="pingpong" />

I think for now we'll simply keep this as-is in BP XML.  So let's auto-gen. the wiring XML for all <bean> but keep <reference> and all <odl:*> as is, would be my suggestion, to get this going.

I'm assuming a bundle EITHER uses the std. OR this ping-pong (whatever it is), right?  Any place with an @Inject DataBroker then gets ping ping balls injected - should work just fine.  If we need BOTH implementations, then we can start playing with @Named - but this probably won't even be needed.

 > Basically it's using a service property ("type") to pick which
 > implementation to get from the OSGi service registry and inject.

Just FYI if in the future we want to move away from XML completely, then it looks like we may be able to use @OsgiService with its filter()for this - but I don't understand exactly how either. -- The bigger fish to fry before worrying about that would probably be how to create Java APIs for all the (your!) <odl:*> BP.  But again, I'd suggest that we tackle all of that later, if ever - IMHO it's very valuable to get even just the wiring XML for all "plain" <bean> generated already.

Patch-set: 4