Update patch set 1
Patch Set 1:

Apparently Neutron allows a subnet creation with "--no-gateway" option. We can also spawn VMs on such subnets. Although there are valid reasons in IPv4 use-cases, in Ipv6 use-cases I don't see any specific use-case.

Anyways, since there is a possibility for a user to create an Ipv6 subnet with "--no-gateway" option, we have to handle this use-case.

I verified that after creating a V6 subnet without gateway, we are able to spawn VMs, but we will not be advertising any RAs to such VMs.
Verified couple of use-cases and they all are working fine.

BTW, Neutron does not allow a V6 subnet without gateway to be associated to a Router. So, we are covered.

Patch-set: 1
Reviewer: Gerrit User 5197 <5197@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0