Update patch set 1
Patch Set 1:

Apparently if the the subnet is added with no-gateway option, then i suppose the null check has to added in all retrieval places. Though Neutron does not allow a router to be associated with v6subnet without gateway so we are safe now. If/is there a possibility of this association available in future.

Patch-set: 1