Update patch set 1
Patch Set 1:

Reg: "If/is there a possibility of this association available in future." - Atleast not in the pipeline and may not be in the immediate future. Even if its done, Neutron would expect the user to update the subnet with the gateway first before associating the subnet to the Router.

Patch-set: 1