Fix DMAC flows for Vlan E/W provider
This fix addresses two bugs related only in that
they both prevent vlan provider networks from
working:

1) Wrong (remote) dmac flows generated for Vlan provider networks
https://bugs.opendaylight.org/show_bug.cgi?id=6583

2) Mac learning in vlan provider networks clobbers static flows with wrong flows
https://bugs.opendaylight.org/show_bug.cgi?id=6584

The fix for the (1) is to test if the net is
of type vlan and then retrieve the correct action.

The fix for (2) is to simply disable configuration
of remote dmac flows when the interface from whence
entered the unknown smac is a vlan or flat network
provider interface.

Change-Id: Id3431e94d96dac5230679fcb70fe0a79efbd96a4
Signed-off-by: Josh <jhershbe@redhat.com>