Fix DMAC flows for Vlan E/W provider DO NOT MERGE
This fix addresses two bugs:
https://bugs.opendaylight.org/show_bug.cgi?id=6583
https://bugs.opendaylight.org/show_bug.cgi?id=6584

The fix for the first bug is to test if the net is
of type vlan and then retrieve the correct action.

The fix for the second is to simply disable configuration
of remote dmac flows. There is currently some conversation
as whether this is the correct approach so this fix
should not be pushed until that is resolved

Change-Id: Id3431e94d96dac5230679fcb70fe0a79efbd96a4
Signed-off-by: Josh <jhershbe@redhat.com>