Fix DMAC flows for Vlan E/W provider DO NOT MERGE
This is a first cut at a solution. It works...
...sort of. It creates the correct flows however
there is a separate issue where the flows are
deleted after some time. However, if you manage
to test it before then, everything works ;-)
I want to (a) clean up this code a little in an
additional patch set to come and (b) resolve the
deletion issue before this patch is merged.

Change-Id: Id3431e94d96dac5230679fcb70fe0a79efbd96a4
Signed-off-by: Josh <jhershbe@redhat.com>