Fixes for import/export RT and router dissociation in L3Vpn
A label of 0 was being used to create unique ids across vpn engine if
IdManager failed to return a valid label.
Java Synchronize has been replaced by LockManager to handle events
gracefully across PLs during simultaneous northbound and southbound
events for subnet routes.
To find out all the vpns who import routes of a particular vpn, we were
dependent on the configurational datastore of that Vpn Instance. On a vpn
instance deletion, this config data is lost. Hence, this data has to also
be stored in an operational datastore

Change-Id: Idbf3d2c6be6a83705043dc76cc14935e02730f26
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>