Update patch set 9
Patch Set 9:

FYI everyone I've locally run NetvirtIT with this change, and it (still) passes, so IMHO this should now be low risk to merge. -- The only real "danger" I can imagine would be if anything was merged in the last 48h (since I finished this last Friday) which changed the wiring (again?!) - with any new beans or changes to the BP XML possibly need this to be adapted, again.  But as there are no merge conflicts, it should be just fine IMHO.

Patch-set: 9