Reduce IPv6 Periodic Unsolicited RA timer interval
In some situations it was noticed that flow programming is taking
some time. Due to this, the Router Soliciation requests from the
VMs are not punted to the controller (VMs are able to acquire IPv6
address only upon arrival of unsolicited Router Advts from the ODL
controller). Currently the periodic unsolicited RA interval is set
to 5 mins. This patch reduces the interval to 60 seconds to mitgate
such situations.

In a future patch, we would enhance the periodic RA interval to an
updated algorithm which will provide a better user experience.

Change-Id: I2eb4061b487cfa709ef02b7d093481799294adb3
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>