Support multiple routers per external GW
Was finally able to see the basic tempest scenario test
passing with this patch! (with a small tempest workaround
to disable ping to router interface, which is currently not
supported in netvirt)

It includes the following changes:
 * Create a vpninstance (with a new vpnId) for the neutron external network
 * Associate the external interface with that vpninstance
 * Partially revert changes done in [1] & [2], to make sure external
   floating IP & NAT rules are associated with the new vpnId
 * Fix ExternalNetworkGroupInstaller to look for the gateway IP in the
   under the new vpnId. Also do some refactoring to allow better debug messages
   for ARP resolving.
   Also, drop traffic sent to external gw traffic if its mac is not resolved (Bug 6653)

This code leverages current implementation of NAT tables to translate external vpnId
to internal vpnIds for specific floating / PNAT IPs.

[1] https://git.opendaylight.org/gerrit/#/c/43303/
[2] https://git.opendaylight.org/gerrit/#/c/43102/

Change-Id: I8d3019389d7fe6fc923a509082e7a13d61a91444
Signed-off-by: Koby Aizer <koby.aizer@hpe.com>