Update patch set 2
Patch Set 2: Code-Review+1

> Michael, cold you rebase 

Done.  Note that as part of the rebased patch set 2, I've gone a little further than initially, and also cleaned up a handful on other (non-Exception related) remaining/new Checkstyle violations.  With this, as of this commit, there are now no longer any.  Therefore, I suggest to use this as the opportunity to switch over to enforcing it, and fail the build for changes which introduce new violations, is that OK at this point?  Beware, this could cause other recent gerrits which may introduce new checkstyle violations to fail the build! I would therefore rebase anything that's pending after this has been merged.

> cherry-pick to stable/boron? Then we can merge this.

Will do when build passed.

Patch-set: 2
Label: Code-Review=+1