Update patch set 2
Patch Set 2:

Sridhar, 
The issue that you mention relates to the other commit. In the log you should see that genius fails to add this bridge and therefore the tunnels could not be created. 
In the other review I added a listener to a change in the BridgeRefEntity to try and recreate the tunnels when the state of the bridge recovers. It seems that in your case it didn't. 
If this is the case then it is a different issue and we should create a different bug for that.

Patch-set: 2
Reviewer: Gerrit User 3511 <3511@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0