Fix bad design of AclClusterUtil to make it pluggable for e2e tests
Having AclClusterUtil as a "Bean" with a @PostContruct start() method
but then a static method (isEntityOwner()) which relies on that start()
is bad design - it's either static or it's not, not half-half.  It's
better not be static, and have an interface; this way the (upcoming)
end-to-end AclServiceTest can easily plug in an "always true"
implementation.

Change-Id: Ia2c07ba1cfc5b7447fcd10ad88e1008fca3df35f
Signed-off-by: Michael Vorburger <vorburger@redhat.com>
(cherry picked from commit cc82e251a77ad76659036a9bb57516539dd17cef)