Bug 6732: ARP Replies Intermittent for Floating IP Addresses
Don't stop ARP responder logic in case there is an ARP cache entry
for targetIp.
Even though this cache entry should never exist for IPs that the ARP responder
is responsible for, there is no reason to stop the ARP request
processing because of that.

Disabled AclServiceTest#newInterface() because it was failing.
Will fix it in another patch since it has nothing to do with this patch.

Change-Id: I392d9d9132350201e8d5761fe99c348223ea858d
Signed-off-by: Koby Aizer <koby.aizer@hpe.com>
Signed-off-by: Sam Hague <shague@redhat.com>