Initial changes in the NeutronVPN, FIB and VPN Engine
Added data model changes in the NeutronVPN, VPN and FIB yang files. Network-Id is now
included in the VpnIntance configuration model.
NeutronVpn changes to include VPN type in the VPN instance config data.
Type added to handle requests from both neutron northbound and CLI
command. Default L3 Type asoicated with Internal VPN
VpnInstance Opertional data is now updated with the appropriate VPN Type.
FibEntries are written with either the label(For L3), or with the vni, VM
mac and the gatewayMac(For L2).

Change-Id: I748901a96545466c5e64e214a7ae528b75264be7
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>