Initial changes in the VPN Engine
Added data model changes in the VPN and FIB yang files. Network-Id is now
included in the VpnIntance cpnfiguration model.
VpnInstance OPertional data is now updated with the approprate VPN Type.
FibEntries are written with either the label(For L3), or with the vni, VM
mac and the gatewayMac(For L2).

Change-Id: I748901a96545466c5e64e214a7ae528b75264be7
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>