Upgrade AssertBeans version, intro. nicer syntax w.o. build()
Also incl. some general clean-up: now all code in *-impl instead of some
in *-api, as that is probably much cleaner, and simpler to follow when
reading.

The 3 classes that are in test.infra will eventually move somewhere else
(but let's anyway get this in as proposed here ASAP already).

Change-Id: Iaf136a88d1f6740a56159efc3c5349348b9516b0
Signed-off-by: Michael Vorburger <vorburger@redhat.com>