Update patch set 2
Patch Set 2: Code-Review+1

Yeah this should fix the bug in question.
Aswin I didn't totally understand the following, but I trust you understand it better than me :)

a. In what case would we need a rule with "drop" behavior for SG? If it wouldn't hit this rule, it wouldn't hit any rule and then it would be dropped by default, wouldn't it?
Referring to rules like this:
 cookie=0x6900000, duration=17908.260s, table=40, n_packets=0, n_bytes=0, priority=63010,udp,metadata=0x2b5a0000000000/0x1fffff0000000000,tp_src=67,tp_dst=68 actions=drop
 cookie=0x6900000, duration=17908.258s, table=40, n_packets=0, n_bytes=0, priority=63010,udp6,metadata=0x2b5a0000000000/0x1fffff0000000000,tp_src=547,tp_dst=546 actions=drop
 cookie=0x6900000, duration=17908.253s, table=40, n_packets=0, n_bytes=0, priority=63020,icmp6,metadata=0x2b5a0000000000/0x1fffff0000000000,icmp_type=134,icmp_code=0 actions=drop

b. Why are transparent security groups different from the other SG in terms of not using isOfInterest? For all implementations if port security is disabled we should just skip the ACL, and if it is we should enter the ACL tables with whatever logic is in there (default allow for transparent and default drop for other two), no?

Patch-set: 2
Reviewer: Gerrit User 670 <670@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1