Update patch set 4
Patch Set 4: Code-Review+1

Hi Sam / Vishal,

I unit tested these changes with the following use -cases:

a. Created a router and assigned 2 subnets to that router
b. Make sure VMs on that router are able to communicate with each other.
c. Created an L3VPN and then moved a router in to it.
d. Made sure that subnetroute entry appeared and also made sure that VMs are able to communicate with each other.

Patch-set: 4
Label: Code-Review=+1