Bug 6721: first few ping requests to a floating IP are receiving multiple responses
This fix attempts to prevent multiple Datapaths replicating
an IP response from external network to floating ip hosted
inside the cloud.

As part of this fix, we install floatingipportmac into
MyMac Table (Table 19 at the time of this commit),
only on those hypervisors where that floating-ip hosted
VM is present.

Please note that the duplicate IP response packets
replication problem does not exists when external network
uses tunnels (instead of VLANs) as the floatingip destined
response packets from external network will be directed
to only one hypervisor (ie., only one datapath).

The fix also makes sure that when a floatingip is
disassociated the corresponding floatingipportmac entry
is removed from MyMac Table.

There are still general bugs where floatingip rules in some
tables (including MyMac table) are not removed when:

a. Floating IP is deleted
b. VM hosting the Floating IP is deleted (without disassociating).

The workaround for both the above is to disassociate floatingip
first before embarking on a (or) b . Issue a and b will
be addressed with a different bug, as it needs a more deeper
fix.

Change-Id: Ic0eb7994a9829aa89aee5d8400e0d0be94dfac36
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>