Ethernet VPN Enhancement to NeutronVpn.
This review brings in data-model enahancements and code changes in
neutronvpn to support BGP L3VPN support over VxLAN in OpenDaylight. This is a
first in a series of reviews that will have 'EVPN_RT5' as the topic.

This review bring in the following things :
1) Data model changes for config vpn-instances to support VPN type and
l3vni
2) l3vni attribute added to each vpn instance.
3) Code changes to service instance and interface creation for L3VPN over
VxLAN in addition to the existing L3VPN over MPLSoGRE.

Change-Id: Ic46cfb8886eb429172151f8f0239a0166468e3ca
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>