Ethernet VPN Enhancement to NeutronVpn.
This review brings in code changes in neutronvpn to support
BGP L3VPN support over VxLAN in OpenDaylight. This is the second review
with 'EVPN_RT5' as the topic.

This review bring in the following things :
1) Code changes to service instance and interface creation for L3VPN over
VxLAN in addition to the existing L3VPN over MPLSoGRE.
2) Minor yang change to use a generic VpnInstance data-model in
NeutronVpn, when handling vpn instances through RPC

Change-Id: Ic46cfb8886eb429172151f8f0239a0166468e3ca
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>