Ethernet VPN Enhancement to FIB Engine
This review brings in data model enhancements and code changes in FIB
Engine to support BGP Ethernet VPN feature in OpenDaylight.  This is the
third review in the 'EVPN_RT5' topic branch.

This review brings in the following :
1) Data model and API changes to add vni, mac address and gateway mac address to a
Vrf Entry representing a route for an EVPN interface.
2) Data model changes to set the encapsulation type (mplsgre/vxlan) for a
route based on its association to an EVPN or L3VPN.

Change-Id: Ica66e2c78f8e1fc4055520b578a8260f13d1c144
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>