Ethernet VPN Enhancement to Vpn Engine and FIB.
This review brings in data model enhancements and code changes in VPN
Engine and FIB to support BGP Ethernet VPN feature in OpenDaylight.
This is the second review in the 'EVPN_RT5' topic branch.

This review brings in the following :
1) Data model changes to specify the type of VPN being processed.
2) Code changes to add FIB entries and advertising routes to BGP,
appropriate to the type of VPN being processed.
3) Data model and API changes to add vni, mac address and gateway mac
address to a Vrf Entry representing a route for an EVPN interface.
4) Data model changes to set the encapsulation type (mplsgre/vxlan) for a
route based on its association to an EVPN or L3VPN.

Conflicts:
	vpnservice/fibmanager/fibmanager-api/src/main/java/org/opendaylight/netvirt/fibmanager/api/IFibManager.java
	vpnservice/fibmanager/fibmanager-impl/src/main/java/org/opendaylight/netvirt/fibmanager/FibUtil.java
	vpnservice/vpnmanager/vpnmanager-impl/src/main/java/org/opendaylight/netvirt/vpnmanager/VpnInterfaceManager.java
	vpnservice/vpnmanager/vpnmanager-impl/src/main/java/org/opendaylight/netvirt/vpnmanager/VpnUtil.java

Change-Id: I92ec69843151b93720a53c17436878707aa0a2b4
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>