Update patch set 6
Patch Set 6:

I think the reverse traffic will stop working with this patch in stateless mode and is better to fix it. 
Now we require to add rules in both direction for traffic to be allowed which is not expected.

Patch-set: 6
Reviewer: Gerrit User 4269 <4269@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0