Bug 6831: Retain subnetroute with l3 directly-connected subnet
The following review :

https://git.opendaylight.org/gerrit/#/c/46292/

was reverted via:
https://git.opendaylight.org/gerrit/#/c/46575/

The revert was done in order to maintain sanity of the
subnetroute feature which enables auto-discovery of
Non-Neutron IP addresses on an enterprise network by using
the datacenter as the proxy.

The revert however resulted in use-case breakage where rewrite of
MACAddresses held behind a directly connected subnet ceased
to happen.

With this fix, we bring back the use-case of rewrite of mac-addresses
for hosts behind a directly connected l3 subnet.  In addition to that
we attempt to maintain a working subnetroute functionality.

Change-Id: I89a61396d62e38cf9756cbd0bdb2b845b5e8e44f
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>