Bug 6853 : Directly removing floatingip doesn't rmv mac entries from T19
NAT model change to accomodate and pass the mac address of the floating ip
port for creating table 19 entries, obtained via the created neutron port
upon a floating ip create.

Added some TODOs as pointers which may help in subsequent review by Hari
Krishna to change fxn signatures to pass on the mac to NAT workflow.
Retry logic for fetching the mac in case of failures due to racing will be
added in the next PS.

Added floatingip-delete T19 fix to remove mac-address

Change-Id: I0bd8fd3937c91e885e77aeaafe4522ce4751b0fa
Signed-off-by: Abhinav Gupta <abhinav.gupta@ericsson.com>