Bug 6904 : ELANInstance read causing NPE on cluster reboot while sending notifs
ELANInstance read was causing NPE on cluster reboot while sending notifs,
leading to breakage in subnetRoute feature.

This commit depends on [1] and addresses the following:

1. On Cluster Reboot, the DataStore events are getting triggered out of
order and hence some data are not available when processing subnet and
ports addition. This fix addresses fetching elanTag from elanDatastore.
If the elan datastore is not written with the elanTag, then the elanTag
is tried to be retrieved from the IdManager.

2. Added logs indicating creation/deletion of ELAN instance.

3. Improved try-catch error handling 4. Proceeding with DjC transaction
for other operations in case this fails.

[1] https://git.opendaylight.org/gerrit/#/c/47131/

Change-Id: Icb09f4a6081b18a01ed6dc4590e695cf04fb9eb5
Signed-off-by: Abhinav Gupta <abhinav.gupta@ericsson.com>