BUG 6940 - Avoid TZ subnet per neutron subnet
Before this patch, each neutron subnet was being mapped to a transport
zone subnet. This is not correct, as the transport zone subnets relate
to the provider subnet each TEP belongs to and not to overlay subnets.
The observed problem is that the same TEP could be configured multiple
times on different transport zone subnets. This would result on ITM
getting confused and thinking that some TEPs are not longer needed and
removing them.

Since at this point there is not particular reason to group the TEPs by
subnets, this change proposes to configure the TEPs on the 0.0.0.0/0
subnet.

Change-Id: I3897f8c03a7a966fe97f4215c0d3ed4388167f50
Signed-off-by: Jaime Caamaño Ruiz <jaime.caamano.ruiz@ericsson.com>