Update patch set 8
Patch Set 8:

Hi Alon,

The 90 seconds is the maximum waiting time to see if a VM available in the router interface can be plumbed successfully.. Since we have a synchronized there, the wait is usually in microseconds for normal scenarios.

The 90 seconds wait time is hit only during the scenario wwhere we create about 200 Routers (i.e, VPNInstances) each with 10 VMs in one of its interfaces. Under such condition, the VPNManager does not get the CPU time to create the RouterInstance while the VPNInterfaceManager puts its hands on to try and plumb the VM interfaces on that Router.

A subsequent review is being worked on and will land in carbon where we will put the vpninterfaces into a pending queue and service them as soon as the parent vpn instance is available instead of waiting with fixed delays (which will usually break as you further increase the number of Routers (or) VMs in a router).

This review is clearly a stop gap and so I will raise a bug to remove this waiting cycles.  That will make sure we track the removal of waiting times from these files as part of the subsequent review.

Patch-set: 8