Update patch set 1
Patch Set 1:

As we discussed , you can pass the VRFEntry IPPrefix as an additional parameter to createLocalNextHop() and removeLocalNextHop().
You need to retrieve the adjacency for that VRFEntry IPPrefix and check if a MACAddress is available for the same.
If it is available use that as an additional key for VPNNextHop Datastore.. 
If a MACAddress is not available, use the MACAddress from the adjacency 
retrieved from localNextHopIp (passed in as ipAddress to createLocalNextHop() already).  

Based on that key create a new group (or) use an existing group.  
In the same way delete the group before the adjacency is removed.

Please make this review a dependent on my review 

https://git.opendaylight.org/gerrit/#/c/46618/

and repost the next PS.

Patch-set: 1