Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

looks like the csit failed badly - all around the ping test, which is usually a sign a flow (or flows) were not installed. The changes for the most part look OK though. Is there a possibility the datatree is slightly different in the events that are received and thus a flow is not being installed?

Patch-set: 6
Label: Code-Review=-1