Update patch set 19
Patch Set 19:

> (1 comment)
 > 
 > The only thing is clean up of external group entry does not seem to
 > happen when a DPN exits from an Internal VPN..
 > 
 > Other than that these changes look fine..
 > 
 > Request you Sam to decide whether to go forward with the merge :)
 > 
 > Vivek

Hi Vivek,

As a single ext-net may be associated with multiple routers, the group entry for each subnet within the ext-net needs to be removed only when there are no more associations to any router and not to a single router. As ext-net router associations may be updated concurrently we decided it would be safer to do the ext-net group cleanup for each subnet only when the subnet itself is removed only because trying to figure out whether this is the last router associated to the ext-net seems like it can be more error-prone.

See org.opendaylight.netvirt.natservice.internal.SubnetmapListener.remove() triggers ext-net group removal on subnet removal.

I could add a code to check odl-nat/external-networks model in RouterDpnChangeListener.remove and see if the router list contains only the current router-id. It just seems like it could lead to race conditions and we may end up with no installed group entry. Since this is just for cleanup purposes I prefer to leave the code as is

what do you think?

Patch-set: 19
Reviewer: Gerrit User 5802 <5802@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0