Bug 7020: Deletion issue when VM has multiple SGs with same rules
Issue: When VM has multiple SGs with same rules, single flow is created on
switch for the common rules.
When one of the SG is dissociated from the VM, that particular flow is
deleted from the switch even though there is another SG with same rules.

Fix:
- For all overlapping rules, individual flows are created on the switch
  with different priorities.
- Per SG, unique flow priority is created using IdManager. Pool range is
    1000-60000.
- Flow priorities are cached to avoid multiple lookups.
- Note, CT_STATE_NEW_PRIORITY_DROP is now set to value 50

Fix provided for stateful mode only.

Change-Id: I1d3c69883aeadb786536c3e7bf76534c1f137336
Signed-off-by: Somashekar Byrappa <somashekar.b@altencalsoftlabs.com>