[WIP] IPv6 East-West Routing Support
This patch supports IPv6 east-west communication using OpenFlow rules.

Sample setup:
-------------
Create a Neutron router with two networks (say network1 and network2).
Create an IPv6 subnet (or dual-stack) in each network.
Associate all the subnets to the router.
Spawn VMs on network1 and network2.
Verify that ping6 from VM1 (on network1) to VM2 (on network2) is successful.
ODL virtual router should be able to support routing between the two networks
using OpenFlow rules.

Note:
Currently this patch is verified for an all-in-one node with transparent
SG (i.e., Security Groups disabled).

Todo:
-----
Have to verify the use-case on a multi-node setup and with Stateful SG enabled.

Change-Id: I7dff4dcccf9e746656ca04cb09f99b31f585fcd8
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>