Bug 7024: When router is associated to L3VPN , VRF entry creations takes long time
When router is associated to L3VPN and dis-associated it takes long time
for VRF entries to get created and also for table=21 entries to appear on
DPN.

When swap from internal-to-external VPN or vice-versa happens,
VpnInterfaceManager waits on notify task for 50sec eventhough FIB has
already cleaned up. This is due to notification not fired from VpnInterface
Oplistner.

Change-Id: Ic16febd1f322f5c79d17e428a8f95133e7fef227
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>