Update patch set 10
Patch Set 10:

Aswin, I think it helps but does not solve all the timing issues. The first timing issue is with port creation.

The 30s sleep seems to fix that issue - but that has been there a while.

The next is tunnel ports. The ports took a while to come up so I added the wait for them. But there are still cases where they just don't get created. That is the only time I see the failures in the single test run.

In the full run of all tests I saw the ovsdb plugin hit a NPE and that failed a single test. I filed a bug for that. Pretty sure it has nothing to do with NetVirt.

Patch-set: 10
Label: Verified=0