Update patch set 2
Patch Set 2:

(1 comment)

For me it is easier to use the "flow:1", since the dpnId is directly retrievable from the nodeId. Also I'd guess that it'll fire even less DCNs than the "ovsdb:1".

Unless Shuva confirms that there could be some issue with "flow:1", I'd stick to it.

Having said that, it is not too much trouble to switch to "ovsdb:1"

Patch-set: 2
Label: Verified=0