Update patch set 3
Patch Set 3:

(2 comments)

I think it is not optimal way of solving this problem.
I think better way would be to use reference-counters for rules.
For example to use mapping of rules (or matches of the flows) to Integer, to increase it for same rules and to decrease it for removed rule. The flow will be removed if we hit zero for specific rule.
In this way there will be no flows which might be never used because their low priority.
Is that make sense?

Patch-set: 3
Reviewer: Gerrit User 6238 <6238@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0