Bug 6786: L3VPN is not honoring VTEP add or delete in operational cloud
As part of making L3VPN honor VTEP add-delete operation , this is a second
set of changes.

This code review contains a Tunnel state listner which listens on VTEP
add/remove events. When VTEP add event is received , interface manager
is queried for list of VpnInterfaces for a DPN on which VTEP is added.
For all those VpnInterfaces , adjacency is updated with a valid VTEP IP as
nexthop IP and also corresponding VRF entry is updated. Similarly on
receiving VTEP delete event , VpnInterface adjacency and VRF entry is
updated.
Also flow rules on DPN are updated accordingly on DPN.

Change-Id: Idca8e0ea674e1c9cfee2ba97b06ed1ae8b17297e
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>
Signed-off-by: ehvkand <hanamantagoud.v.kandagal@ericsson.com>