Bug 7091 : When Primary NAPT switch goes down, NAPT switch re-election is not happening.
The root cause of this issue was the change in the representation of the
ietf-interfaces:interfaces-state model.

Since NAT models depends upon the ietf-interfaces:interfaces-state model,
NAT HA functionality of re-electing the NAPT switch never got triggered
and the functionality got broken. As per the new representation of the
'ietf-interfaces:interfaces-state' model, 'oper-status' attribute value is
'unknown' if the corressponding DPN goes down.

update() of the NatInterfaceStateChangeListner is implemented to handle this change for
clearing or populating the 'dpn-vpninterfaces-list' in the NAT model
'odl-l3vpn:neutron-router-dpns' used by the NAT HA code.

Change-Id: I533c9c1c5d32ab5967cd21e7b0f203978b074703
Signed-off-by: Satish Dutt <satishd@altencalsoftlabs.com>