Update patch set 3
Patch Set 3:

'odl-l3vpn:neutron-router-dpns' is the main model which is being used by SNAT HA code. This model reflects the OPERATIONAL SWITCHES for a router. So if the NON-OPERATIONAL SWITCHES (Eg. SWITCHES GOING DOWN) are not removed from this model, then there will be an INCONSISTENCY in the switches which provides the SNAT service. For example, if a NAPT switch goes down, then the corresponding switch should be removed from the model 'odl-l3vpn:neutron-router-dpns' and the SNAT HA code which listens on this model will re-elect a new NAPT switch and configure the SNAT flows on the newly elected NAPT switch. 

Even if the bug https://bugs.opendaylight.org/show_bug.cgi?id=7081 mentioned by you is resolved, you will still see the SNAT HA functionality broken due to the above reason and have a new bug https://bugs.opendaylight.org/show_bug.cgi?id=7091 to track this issue

Patch-set: 3