ARP resolution fix for VPN
Current ARP Handling code was trying to form
learned vpnportiptoport entries for configured
neutron ports incorrectly.  This is observed to
be happening primarily during the swap window from
internal to external VPN and vice-versa.

The fix attempts to make sure that the VPNInterface
we are servicing ARP for is indeed the same as configured by the
tenant.

Change-Id: I17bb53a2997e77a9ba7ce13284525ac403fd2bac
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>