Update patch set 2
Patch Set 2:

The main intent of this is to get rid of those horrible exceptions we hit in the bug. I assume this was not meant for non-neutron ports as they obviously don't have a gateway.

I am not sure if there is a better method to do this check without using ElanUtils - let me know if something comes to mind.

Patch-set: 2
Label: Verified=0