Ethernet VPN : Data model enhancement to add l3vni
This review brings in the data model changes in VPN, FIB and NeutronVPN
engines, introducing the l3vni attribute. Either an RPC or a direct VPN
Instance object can be written into the DataStore to create a VPN of Type
L2.
For a VPN of type L2, the l3vni input attribute is mandatory.

Change-Id: I7e194e155f6a7b505206b702caa93a4f3dde46b4
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>