Update patch set 16
Patch Set 16: Code-Review-1

> (5 comments)

Hi Karthik,

We usually don't comment code - if it's not supposed to be there anymore, delete it.
Also please update the commit message - this is not an ARP responder for router gateways - as I understand it this commit is supposed to be an ARP responder for router interfaces AND floating IPs (please add if missing). 
Router gateways will be added by Tali once this is committed as agreed with Vivek, as they have some more nuances.

In addition, it would be great if you could get 'test-netvirt-cluster-current' passing with this commit, the ARP responder is supposed to fix the behavior under failure. Vishal can probably assist if required.

On a related topic, I'm a little bit worried about continuing to punt all ARPs to the controller, specifically when talking about ARPs arriving from the external network. We were thinking of only punting the ARP from the master (NAPT?) switch in this case. Tali might be able to do this later as part of the router gateway commit (assuming SH bit is set).

Patch-set: 16
Label: Code-Review=-1